+++ app/src/main/java/com/example/blognhom2/Fragment/PostPendingContentFragment.kt
@@ -22,40 +22,20 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import java.sql.DriverManager
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [PostPendingContent.newInstance] factory method to
- * create an instance of this fragment.
- */
 class PostPendingContentFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    lateinit var post : PostInfo;
+    lateinit var post : PostInfo
     private var _binding: FragmentPostPendingContentBinding? = null
 
 
     private val binding get() = _binding!!
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         _binding = FragmentPostPendingContentBinding.inflate(inflater, container, false)
-
+        SetDataForPostContent()
         binding.cancelBtn.setOnClickListener {
             cancelPost()
             showDialog()
