+++ app/src/main/java/com/example/blognhom2/LoginActivity.kt
@@ -18,6 +18,7 @@ import retrofit2.*
 import retrofit2.converter.gson.GsonConverterFactory
 import java.sql.DriverManager.println
 
+var userRole: String?= null
 class LoginActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityLoginBinding
@@ -28,16 +29,16 @@ class LoginActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         binding.loginButton.setOnClickListener {
-//            val email = binding.loginEmail.text.toString()
-//            val password = binding.loginPassword.text.toString()
-//            if (email.isNotEmpty() && password.isNotEmpty()) {
-//                prepare(email, password)
-//            } else {
-//                Toast.makeText(this, "Fields cannot be empty", Toast.LENGTH_SHORT).show()
-//            }
-            val mainIntent = Intent(this@LoginActivity, MainActivity::class.java)
-            startActivity(mainIntent)
-            finish() // Close the LoginActivity
+            val email = binding.loginEmail.text.toString()
+            val password = binding.loginPassword.text.toString()
+            if (email.isNotEmpty() && password.isNotEmpty()) {
+                prepare(email, password)
+            } else {
+                Toast.makeText(this, "Fields cannot be empty", Toast.LENGTH_SHORT).show()
+            }
+//            val mainIntent = Intent(this@LoginActivity, MainActivity::class.java)
+//            startActivity(mainIntent)
+//            finish() // Close the LoginActivity
         }
 
         binding.forgotPassword.setOnClickListener {
@@ -97,6 +98,9 @@ class LoginActivity : AppCompatActivity() {
                 }
                 val user = response.body()
                 if (user != null) {
+
+                    userRole = user.roles[0]
+
                     Log.d("LoginActivity", "User authenticated: $user")
                     // Handle the authenticated user here
                     val setCookieHeader = response.headers()["Set-Cookie"]
