+++ app/src/main/res/drawable/custom_button.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+
+    <solid android:color="#5EBF8F" />
+    <corners android:radius="10dp" />
+
+</shape>
