+++ app/src/main/res/drawable/spinner_custom.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+<item>
+    <layer-list>
+        <item>
+            <shape android:shape="rectangle">
+                <solid android:color="#E6E3E3" />
+                <corners android:radius="10dp" />
+            </shape>
+        </item>
+    </layer-list>
+</item>
+</selector>
