+++ app/src/main/res/layout/fragment_write.xml
@@ -20,7 +20,7 @@
     <ImageView
         android:id="@+id/wImage"
         android:layout_width="wrap_content"
-        android:layout_height="256dp"
+        android:layout_height="200dp"
         android:layout_marginBottom="4dp"
         android:scaleType="fitXY"
         app:layout_constraintBottom_toTopOf="@+id/wPickImgBtn"
@@ -33,7 +33,10 @@
         android:id="@+id/wPickImgBtn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:background="@drawable/custom_button"
         android:text="Pick Img"
+        android:textColor="#FFFFFF"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
@@ -41,6 +44,7 @@
 
     <Spinner
         android:id="@+id/wCategoies"
+        android:background="@drawable/spinner_custom"
         android:layout_width="match_parent"
         android:layout_height="52dp"
         android:layout_marginTop="4dp"
@@ -66,8 +70,14 @@
         android:id="@+id/wSubmitBtn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="8dp"
+        android:layout_marginBottom="8dp"
+        android:background="@drawable/custom_button"
         android:text="Submit"
+        android:textColor="#FFFFFF"
         app:layout_constraintBottom_toBottomOf="parent"
-        tools:layout_editor_absoluteX="40dp" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
